+++ R/SDistribution_Gompertz.R
@@ -55,19 +55,84 @@ Gompertz$set("public", "pgf", function(z){
   return(NaN)
 })
 
-Gompertz$set("private",".pdf", function(x){
-  self$getParameterValue("shape") * self$getParameterValue("scale")*exp(self$getParameterValue("shape")) *
-    exp(self$getParameterValue("scale") * x) * exp(-self$getParameterValue("shape") * exp(self$getParameterValue("scale") * x))
+Gompertz$set("private",".pdf", function(x, log = FALSE){
+  if (checkmate::testList(self$getParameterValue("shape"))) {
+    mapply(
+      extraDistr::dgompertz,
+      b = self$getParameterValue("shape"),
+      a = self$getParameterValue("scale"),
+      MoreArgs = list(x = x, log = log)
+    )
+  } else {
+    extraDistr::dgompertz(
+      x,
+      b = self$getParameterValue("shape"),
+      a = self$getParameterValue("scale"),
+      log = log
+    )
+  }
 })
-Gompertz$set("private",".cdf", function(x){
-  1 - exp(-self$getParameterValue("shape")*(exp(self$getParameterValue("scale")*x)-1))
+Gompertz$set("private",".cdf", function(x, lower.tail = TRUE, log.p = TRUE){
+  if (checkmate::testList(self$getParameterValue("shape"))) {
+    mapply(
+      extraDistr::pgompertz,
+      b = self$getParameterValue("shape"),
+      a = self$getParameterValue("scale"),
+      MoreArgs = list(
+        q = x,
+        lower.tail = lower.tail,
+        log.p = log.p
+      )
+    )
+  } else {
+    extraDistr::pgompertz(
+      x,
+      b = self$getParameterValue("shape"),
+      a = self$getParameterValue("scale"),
+      lower.tail = lower.tail,
+      log.p = log.p
+    )
+  }
 })
-Gompertz$set("private",".quantile", function(p){
-  log(1 + (log(1-p)/-self$getParameterValue("shape")))/self$getParameterValue("scale")
+Gompertz$set("private",".quantile", function(p, lower.tail = TRUE, log.p = TRUE){
+  if (checkmate::testList(self$getParameterValue("shape"))) {
+    mapply(
+      extraDistr::qgompertz,
+      b = self$getParameterValue("shape"),
+      a = self$getParameterValue("scale"),
+      MoreArgs = list(
+        p = p,
+        lower.tail = lower.tail,
+        log.p = log.p
+      )
+    )
+  } else {
+    extraDistr::qgompertz(
+      p,
+      b = self$getParameterValue("shape"),
+      a = self$getParameterValue("scale"),
+      lower.tail = lower.tail,
+      log.p = log.p
+    )
+  }
 })
 Gompertz$set("private",".rand", function(n){
-  self$quantile(runif(n))
+  if (checkmate::testList(self$getParameterValue("shape"))) {
+    mapply(
+      extraDistr::rgompertz,
+      b = self$getParameterValue("shape"),
+      a = self$getParameterValue("scale"),
+      MoreArgs = list(n = n)
+    )
+  } else {
+    extraDistr::rgompertz(
+      n,
+      b = self$getParameterValue("shape"),
+      a = self$getParameterValue("scale")
+    )
+  }
 })
+Gompertz$set("private",".log", TRUE)
 
 Gompertz$set("public","initialize",function(shape = 1, scale = 1, decorators = NULL, verbose = FALSE){